refactor: convert to loginschema hook
